Bug 1530573 - Protect maybeResizeNursery() from overflow r=sfink
Differential Revision: https://phabricator.services.mozilla.com/D21158

+++ js/src/gc/Nursery.h
@@ -639,8 +639,8 @@ class Nursery {
 
   /* Change the allocable space provided by the nursery. */
   void maybeResizeNursery(JS::GCReason reason);
-  void growAllocableSpace(unsigned newCapacity);
-  void shrinkAllocableSpace(unsigned newCapacity);
+  void growAllocableSpace(size_t newCapacity);
+  void shrinkAllocableSpace(size_t newCapacity);
   void minimizeAllocableSpace();
 
   // Free the chunks starting at firstFreeChunk until the end of the chunks
